move files into source folder
