+++ src/kanjivg/kvg-lookup.py
