+++ src/kanjivg/utils.py
