+++ src/kanjivg/xmlhandler.py
